Mounting the modal inside the Level component
I added a simple div (Box) to the `Level` component, before the "Stake to join" button (with width and height set to 0, so it doesn't break the layout), and I used the reference of that element in `portalProps`. These props are passed down to the `StakingModal`, so it can mount the Chakra UI Modal properly.